Add support for passing build-time environment variables in build context
The build-time environment variables are used to expand the symbols used in certain
Dockerfile primitves like ADD, COPY, USER etc, without an explicit prior definiton using a
ENV primitive. These variables get persisted in the intermediate and final images
whenever they are expanded.

The build-time environment variables are also passed as environment-context for command(s)
run as part of the RUN primitve. These variables are not persisted in intermediate and
final images when passed as context for RUN.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>